Remove dynamics_delta.hpp
